# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fertility Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DailyLogCervicalMucus =
  | (typeof DailyLogCervicalMucus)[keyof typeof DailyLogCervicalMucus]
  | null;

export const DailyLogCervicalMucus = {
  dry: "dry",
  sticky: "sticky",
  creamy: "creamy",
  watery: "watery",
  egg_white: "egg_white",
} as const;

export type DailyLogMood =
  | (typeof DailyLogMood)[keyof typeof DailyLogMood]
  | null;

export const DailyLogMood = {
  great: "great",
  good: "good",
  neutral: "neutral",
  low: "low",
  anxious: "anxious",
} as const;

export type DailyLogPhase =
  | (typeof DailyLogPhase)[keyof typeof DailyLogPhase]
  | null;

export const DailyLogPhase = {
  menstrual: "menstrual",
  follicular: "follicular",
  ovulation: "ovulation",
  luteal: "luteal",
  tww: "tww",
} as const;

export interface DailyLog {
  id: number;
  date: string;
  /** Basal body temperature in Celsius */
  bbt?: number | null;
  cervicalMucus?: DailyLogCervicalMucus;
  mood?: DailyLogMood;
  /**
   * @minimum 1
   * @maximum 10
   */
  energyLevel?: number | null;
  sleepHours?: number | null;
  /**
   * @minimum 1
   * @maximum 10
   */
  stressLevel?: number | null;
  /** List of symptoms (cramps, bloating, spotting, tender_breasts, etc.) */
  symptoms: string[];
  supplements: string[];
  notes?: string | null;
  cycleDay?: number | null;
  phase?: DailyLogPhase;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDailyLogBody {
  date: string;
  bbt?: number | null;
  cervicalMucus?: string | null;
  mood?: string | null;
  energyLevel?: number | null;
  sleepHours?: number | null;
  stressLevel?: number | null;
  symptoms: string[];
  supplements: string[];
  notes?: string | null;
}

export interface Cycle {
  id: number;
  startDate: string;
  endDate?: string | null;
  cycleLength?: number | null;
  /** Estimated cycle day of ovulation */
  ovulationDay?: number | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateCycleBody {
  startDate: string;
  notes?: string | null;
}

export type CurrentCyclePhase =
  (typeof CurrentCyclePhase)[keyof typeof CurrentCyclePhase];

export const CurrentCyclePhase = {
  menstrual: "menstrual",
  follicular: "follicular",
  ovulation: "ovulation",
  luteal: "luteal",
  tww: "tww",
} as const;

export interface CurrentCycle {
  cycle: Cycle;
  cycleDay: number;
  phase: CurrentCyclePhase;
  phaseDescription: string;
  daysUntilNextPhase?: number | null;
  estimatedOvulationDate?: string | null;
  isInTww: boolean;
}

export type GenerateInsightBodyInsightType =
  (typeof GenerateInsightBodyInsightType)[keyof typeof GenerateInsightBodyInsightType];

export const GenerateInsightBodyInsightType = {
  daily: "daily",
  tww_support: "tww_support",
  body_signal: "body_signal",
  energy: "energy",
  boundary: "boundary",
} as const;

export interface GenerateInsightBody {
  logId?: number | null;
  /** Optional additional user context for the insight */
  context?: string;
  insightType: GenerateInsightBodyInsightType;
}

export interface InsightResponse {
  id: number;
  insight: string;
  insightType: string;
  recommendation?: string | null;
  createdAt: string;
}

export interface SavedInsight {
  id: number;
  insight: string;
  insightType: string;
  recommendation?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  currentPhase?: string | null;
  cycleDay?: number | null;
  /**
   * @minimum 0
   * @maximum 100
   */
  readinessScore?: number | null;
  isInTww: boolean;
  recentSymptoms: string[];
  lastLogDate?: string | null;
  totalCycles: number;
  avgCycleLength?: number | null;
  latestInsight?: string | null;
  /** Consecutive days logged */
  streakDays: number;
}

export interface ReadinessScore {
  /**
   * @minimum 0
   * @maximum 100
   */
  overall: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  sleepScore?: number | null;
  /**
   * @minimum 0
   * @maximum 100
   */
  stressScore?: number | null;
  /**
   * @minimum 0
   * @maximum 100
   */
  energyScore?: number | null;
  date: string;
  message: string;
}

export type GetLogsParams = {
  limit?: number;
  offset?: number;
};

export type GetInsightsParams = {
  limit?: number;
};