# lib/api-zod/src/generated/types/createDailyLogBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fertility Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDailyLogBody {
  date: Date;
  bbt?: number | null;
  cervicalMucus?: string | null;
  mood?: string | null;
  energyLevel?: number | null;
  sleepHours?: number | null;
  stressLevel?: number | null;
  symptoms: string[];
  supplements: string[];
  notes?: string | null;
}