# lib/api-zod/src/generated/types/cycle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fertility Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface Cycle {
  id: number;
  startDate: Date;
  endDate?: Date | null;
  cycleLength?: number | null;
  /** Estimated cycle day of ovulation */
  ovulationDay?: number | null;
  notes?: string | null;
  createdAt: Date;
}