# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fertility Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  currentPhase?: string | null;
  cycleDay?: number | null;
  /**
   * @minimum 0
   * @maximum 100
   */
  readinessScore?: number | null;
  isInTww: boolean;
  recentSymptoms: string[];
  lastLogDate?: Date | null;
  totalCycles: number;
  avgCycleLength?: number | null;
  latestInsight?: string | null;
  /** Consecutive days logged */
  streakDays: number;
}