# lib/api-zod/src/generated/types/generateInsightBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fertility Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateInsightBodyInsightType } from "./generateInsightBodyInsightType";

export interface GenerateInsightBody {
  logId?: number | null;
  /** Optional additional user context for the insight */
  context?: string;
  insightType: GenerateInsightBodyInsightType;
}