# lib/api-zod/src/generated/types/insightResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fertility Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface InsightResponse {
  id: number;
  insight: string;
  insightType: string;
  recommendation?: string | null;
  createdAt: Date;
}